pyproject.toml
[project]
name = "gx-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.10.1",
    "albumentations==1.3.1",
    "autopep8==2.0.4",
    "better-exceptions==0.3.3",
    "easydict==1.11",
    "editdistance==0.6.2",
    "flake8==6.1.0",
    "huggingface-hub==0.35.3",
    "hydra-core==1.3.2",
    "imageio==2.33.0",
    "lightning==2.1.3",
    "matplotlib==3.8.2",
    "numba==0.58.1",
    "numpy==1.26.2",
    "opencv-python==4.8.1.78",
    "pandas==2.1.4",
    "pathlib==1.0.1",
    "pillow==10.1.0",
    "polygon3==3.0.9.1",
    "pyclipper==1.3.0.post5",
    "pytorch-lightning==2.1.3",
    "pyyaml==6.0.1",
    "qwen-vl-utils[decord]>=0.0.14",
    "safetensors==0.6.2",
    "scikit-image==0.22.0",
    "scikit-learn==1.3.2",
    "scipy==1.11.4",
    "seaborn==0.13.0",
    "setuptools==69.0.3",
    "shapely==2.0.2",
    "tensorboard==2.15.1",
    "tensorboard-data-server==0.7.2",
    "timm==0.9.12",
    "torchmetrics==1.2.1",
    "tqdm==4.66.1",
    "transformers==4.57.0",
]
